Add Gauge to the name
